+++ cogs/generic_slash.py
@@ -11,6 +11,7 @@ from util.logging.usage import record_usage, finish_usage
 from util.tasks.keep_alive import keep_alive
 from util.tasks.status_change import change_status
 from util.guild_ids import guild_ids
+import util.discord.easy_embed as ezembed
 
 log = convert_logging.get_logging()
 version = get_version()
@@ -123,6 +124,13 @@ class GenericSlash(commands.Cog, description="Generic Functions"):
     @commands.after_invoke(finish_usage)
     async def _version(self, ctx: commands.Context):
         await ctx.respond(f"Bot Version is {version}")
+        
+    @commands.slash_command(guild_ids=guild_ids, name='source', description='Displays Github Source Code')
+    @commands.before_invoke(record_usage)
+    @commands.after_invoke(finish_usage)
+    async def _source(self, ctx):
+        embed=ezembed.create_embed(title='Source Code', description="https://github.com/NottCurious/TMIndiaBot", color=discord.Colour.random())
+        await ctx.respond(embed=embed)
 
 
 def setup(client):